made RefreshTokenController and deletetoken functionality
