+++ src/main/java/com/hsleiden/vdlelie/services/RefreshTokenService.java
@@ -51,7 +51,9 @@ public class RefreshTokenService {
             refreshTokenRepository.delete(token);
             throw new TokenRefreshException(token.getToken(), "Refresh token was expired. Please make a new signin request");
         }
-
         return token;
+    }
+    public void deleteToken(RefreshToken refreshToken){
+        refreshTokenRepository.delete(refreshToken);
     }
 }
